Remove support bot components
